requirements.txt
streamlit
supabase
python-dotenv
pytesseract
Pillow
PyMuPDF
python-pptx
pandas
openpyxl